requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
djongo==1.3.6
pymongo==3.12.1
python-dotenv==1.0.0
requests==2.31.0
GitPython==3.1.40
google-auth==2.25.0
google-auth-oauthlib==1.1.0
google-api-python-client==2.112.0
paypalrestsdk==1.13.1
gunicorn==21.2.0
django-storages==1.14.2
boto3==1.28.64
dj-database-url==2.1.0
whitenoise==6.6.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
django-celery-beat==2.5.0 